requirements.txt
fastapi==0.111.0
groq==0.9.0
langchain==0.2.7
langchain_community==0.2.7
langchain_core==0.2.12
langchain_groq==0.1.6
pydantic==2.8.2
python-dotenv==1.0.1
uvicorn==0.30.1
gpt4all
chromadb
lark
